Preparing the ground to create a validation tool
